Cleaning repo

+++ src/example/metrics.py
@@ -46,7 +46,7 @@ def quat_error(q_des: npt.NDArray[np.float64], q: npt.NDArray[np.float64]) -> np
     return np.arccos(c)
 
 
-def combined_error(ep: float, etheta: float, lambda_theta: float = 1e-6 / np.deg2rad(1) ** 2) -> np.float64:
+def combined_error(ep: np.float64, etheta: np.float64, lambda_theta: np.float64 = 1e-6 / np.deg2rad(1) ** 2) -> np.float64:
     """Compute a weighted combination of position and orientation errors."""
     return np.sqrt(ep**2 + lambda_theta * etheta**2)
